Split the transport emitter and share its wire-binding reads
Two review findings against the inline Rust transport, one mechanical and
one structural:

The emitter had grown past the repo's file-size ceiling. The declaration
text (the bespoke-facing support types and the internal http helpers) moves
to transport_decls.rs, leaving transport.rs with the per-operation emission
alone. The no-reference declarations build from a name/text table instead of
repeated constructor calls, which is also what removes the token-level
duplication the cross-file copy-paste detector flagged between the Rust and
TypeScript emitters (their declaration lists are structurally identical even
where the emitted text differs).

The reads both emitters make off a WireBinding to decide the request shape
(does the operation need a decoded input record, is anything query-bound,
which declared success codes fall outside the 2xx range) are properties of
the binding alone, so they now live once in entries/wire.rs and each target
spells only its own syntax over the answer. TypeScript's emitted output is
byte-identical.

Also from review: resolve_max_retries takes the i64 the frontend's
integer-scalar @retry rule guarantees, dropping the lossy f64 widening and
the is_finite/floor dance at every call site.

+++ examples/payments/sdk/rust/payments/charges/client.rs
@@ -169,7 +169,7 @@ impl Client {
         if !has_header(&headers, "content-type") {
             headers.insert("content-type".to_string(), "application/json".to_string());
         }
-        let max_retries = resolve_max_retries(self.settings.max_retries as f64);
+        let max_retries = resolve_max_retries(self.settings.max_retries as i64);
         let mut attempt: u32 = 0;
         loop {
             let request = HttpRequest {
